+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `full_length_only` parameter to process only full length reads (default: true).
+### Changed
+- Orient 3prime/multiome reads to mRNA sense to avoid need to flip later. 
 
 ## [v1.0.3]
 ### Added